cssom-view/elementFromPoint.html

<!DOCTYPE html>
<title>cssom-view - elementFromPoint</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  .size {
    width:100px;
    height:100px;
  }
  .overlay {
    position:absolute;
    top:109px;
    pointer-events:none;
  }
  .purple {
    background-color: rebeccapurple;
  }
  .red {
    background-color: red;
  }
  .green {
    background-color: green;
  }
</style>
<body>
  <div id='purple' class="size purple" >&nbsp;</div>
  <div id='red' class="size red">&nbsp;</div>
  <div id='green' class="size overlay green">&nbsp;</div>
  <script>
    test(function(){
                  assert_equals(null, document.elementFromPoint(-1, -1),
                    "both co-ordinates passed in are negative so should have returned a null");
                  assert_equals(null, document.elementFromPoint(-1, 1),
                    "x co-ordinates passed in are negative so should have returned a null");
                  assert_equals(null, document.elementFromPoint(1, -1),
                    "y co-ordinates passed in are negative so should have returned a null");
                }, "Negative co-ordinates");
    test(function(){
                  assert_equals(document.getElementById('purple'), document.elementFromPoint(10, 10),
                    "Should have returned the element with id `purple`");
                }, "Return first element that is the target for hit testing");
    test(function(){
                  assert_equals(document.getElementById('red'), document.elementFromPoint(10, 120),
                    "Should have returned the element with id `red` as element with `green` has `pointer-events:none`");
                }, "First element to get mouse events with pointer-events css");
    done()
 </script>
